Permit the linking of MS Teams channels to repo with the LinkRepo command.
This is intended to be a short-term fix while we add support for MS Teams. A better fix might require adding an attribute to the request indicating which chat system is in use. For now we choose to answer the 'isSlack' question by looking at the Channel ID format.

There was already a check to ensure that the channel is public or private (ie. not a DM), which uses the first character of the Slack channel ID. This rejects all MS Teams channels that have a different ID format, so needs changing in order to use MS Teams.